research-agent/cure-for-cancer-agent/requirements.txt
google-generativeai>=0.8.0
requests>=2.31.0
python-dotenv>=1.0.0
xmltodict>=0.13.0
rdkit>=2023.9.1
vina>=1.2.3
google-genai>=1.0.0